Use Jest for tests  - fix babel config  - remove mocha
