--- .babelrc
@@ -1,8 +1,0 @@
-{
-  "presets": [
-    "es2015",
-    "react",
-    "stage-0"
-  ],
-  "plugins": ["transform-class-properties"]
-}
