+++ package.json
@@ -19,30 +19,25 @@
     "postnow-build": "webpack",
     "watch": "webpack -w",
     "dev": "node --experimental-modules \"./src/server/server.js\"",
-    "test": "mocha ./src/__tests__/index.js --require babel-core/register",
-    "test-watch": "mocha ./src/__tests__/index.js --require babel-core/register --watch"
+    "test": "jest"
   },
   "dependencies": {
-    "babel-cli": "^6.23.0",
-    "babel-core": "6.22.1",
-    "babel-eslint": "^7.1.1",
-    "babel-loader": "6.2.10",
-    "babel-plugin-transform-class-properties": "^6.22.0",
-    "babel-preset-es2015": "6.22.0",
-    "babel-preset-react": "6.22.0",
-    "babel-preset-stage-0": "^6.22.0",
     "body-parser": "^1.19.0",
     "copy-webpack-plugin": "^4.0.1",
     "react": "^16.3.0",
     "react-dom": "^16.3.0",
     "react-router-dom": "^4.2.2",
     "express": "^4.17.1",
-    "webpack": "2.2.1",
     "webpack-dev-server": "3.1.11",
     "webpack-node-externals": "^1.5.4",
     "worker-loader": "^0.8.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.10.2",
+    "@babel/preset-env": "^7.10.2",
+    "@babel/preset-react": "^7.10.1",
+    "babel-jest": "^26.0.1",
+    "babel-loader": "^8.1.0",
     "chai": "^3.5.0",
     "cors": "^2.8.1",
     "eslint": "^4.18.2",
@@ -51,8 +46,9 @@
     "eslint-plugin-import": "2.2.0",
     "eslint-plugin-jsx-a11y": "4.0.0",
     "eslint-plugin-react": "6.9.0",
-    "mocha": "^3.2.0",
+    "jest": "^26.0.1",
     "node-fetch": "^2.6.0",
-    "prop-types": "^15.7.2"
+    "prop-types": "^15.7.2",
+    "webpack": "^2.2.1"
   }
 }
