--- src/server/__tests__/findWordsInTrie.js
@@ -1,39 +1,0 @@
-/* eslint-disable arrow-body-style, no-unused-expressions, func-names, prefer-arrow-callback */
-import { expect } from 'chai';
-
-import { findWordsInTrie } from '../lib/findWords';
-import { getTrie } from '../lib/initializeData';
-
-export default function findWordsInTrieTest() {
-  const trie = getTrie();
-
-  describe('findWordsInTrie', function () {
-    it('finds at least one word', function () {
-      const board = 'batter';
-      const foundWords = findWordsInTrie(board);
-
-      expect(foundWords).to.be.an('array');
-
-      const flattenedWords = [].concat(
-        ...Object.keys(foundWords)
-            .filter(wordLength => wordLength >= 3)
-            .sort(wordLength => wordLength)
-            .map(wordLength => foundWords[wordLength]),
-      );
-
-      expect(flattenedWords).to.have.length.of.at.least(1);
-    });
-
-    it('find a valid prefix', function () {
-      // const board = 'abcdeabcdeabcdeabcde';
-
-      expect(trie.isPrefix('b')).to.be.true;
-      expect(trie.isPrefix('ba')).to.be.true;
-      expect(trie.isPrefix('bat')).to.be.true;
-      expect(trie.isPrefix('batt')).to.be.true;
-      expect(trie.isPrefix('batte')).to.be.true;
-      expect(trie.isPrefix('batter')).to.be.true;
-      expect(trie.lookup('battery')).to.be.true;
-    });
-  });
-}
