--- webpack.config.babel.js
@@ -1,3 +1,0 @@
-import ClientConfig from './src/client/webpack.config.client.babel';
-
-export default [].concat(ClientConfig);
